+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `non_mocked_hosts` fixture allowing to avoid mocking requests sent on some specific hosts.
 
 ## [0.7.0] - 2020-08-13
 ### Changed